refactored FWHMTracer
